CHEF-34831: Rename cleanup_vendor_lockfiles.rb to cleanup_gem_lockfiles.rb
Align naming with chef-cli PR #342 convention for consistency
across Chef ecosystem packaging scripts.
