+++ results/last_build.env
@@ -1,0 +1,9 @@
+pkg_origin=ngupta
+pkg_name=cookstyle
+pkg_version=8.7.1
+pkg_release=20260601110230
+pkg_target=x86_64-linux
+pkg_ident=ngupta/cookstyle/8.7.1/20260601110230
+pkg_artifact=ngupta-cookstyle-8.7.1-20260601110230-x86_64-linux.hart
+pkg_sha256sum=0973733e2c86f5357612b27ece2e6da15bf02ae60a6db93f2f04fb8b5886dd2e
+pkg_blake2bsum=4f6260b6d782fb3953a6da293e1d3af2a74dbfddc160dd0963e5c933fce5c4e0
